requirements.txt
networkx ~= 2.4
pyproj ~= 2.4
pyshp ~= 2.1
pyyaml ~= 5.1
requests ~= 2.22
xlrd ~= 1.2
